remove unused code in pt engine
