+++ lmdeploy/pytorch/kernels/__init__.py
@@ -4,7 +4,6 @@ from .alibi_pagedattention import alibi_paged_attention_fwd
 from .apply_rotary_pos_emb import apply_rotary_pos_emb
 from .fill_kv_cache import fill_kv_cache
 from .fused_moe import fused_moe
-from .fused_rotary_emb import fused_rotary_emb
 from .multinomial_sampling import multinomial_sampling
 from .pagedattention import paged_attention_fwd
 from .rms_norm import rms_norm
@@ -14,7 +13,6 @@ from .w8a8_triton_kernels import (matmul_kernel_dynamic_quant, per_channel_quant
 __all__ = [
     'apply_rotary_pos_emb',
     'fused_moe',
-    'fused_rotary_emb',
     'paged_attention_fwd',
     'alibi_paged_attention_fwd',
     'fill_kv_cache',
