+++ lmdeploy/pytorch/kernels/cuda/__init__.py
@@ -7,7 +7,6 @@ from .flash_mla import flash_mla_fwd
 from .flashattention import flash_attention_fwd
 from .flatten_kv_cache import flatten_kv_cache
 from .fused_moe import fused_moe
-from .fused_rotary_emb import fused_rotary_emb
 from .multinomial_sampling import multinomial_sampling
 from .pagedattention import paged_attention_fwd
 from .rms_norm import rms_norm
@@ -17,7 +16,6 @@ from .w8a8_triton_kernels import matmul_kernel_dynamic_quant, per_token_quant_in
 __all__ = [
     'apply_rotary_pos_emb',
     'fused_moe',
-    'fused_rotary_emb',
     'paged_attention_fwd',
     'alibi_paged_attention_fwd',
     'fill_kv_cache',
