+++ lmdeploy/pytorch/kernels/cuda/alibi_pagedattention.py
@@ -8,8 +8,6 @@ import triton
 import triton.language as tl
 from torch import Tensor
 
-from .triton_utils import get_kernel_meta, wrap_jit_func
-
 assert triton.__version__ >= '2.1.0'
 
 LOG2 = tl.constexpr(math.log(2))
@@ -65,7 +63,6 @@ def _load_block_offsets(offset_ptr, block_id, num_sub_blocks: tl.constexpr, BLOC
         return tl.load(offset_ptr + block_id) * BLOCK + offs_n
 
 
-@wrap_jit_func
 @triton.jit
 def _fwd_split_kernel(
     Q,
@@ -200,7 +197,6 @@ def _fwd_split_kernel(
     tl.store(Acc_out + off_meta + 1 + tl.arange(0, 1), l_i)
 
 
-@wrap_jit_func
 @triton.jit
 def _reduce_split_kernel(
     Acc,
@@ -244,7 +240,6 @@ def _reduce_split_kernel(
     tl.store(Out + out_offs, acc)
 
 
-@wrap_jit_func
 @triton.jit
 def _fwd_kernel(
     Q,
@@ -375,7 +370,6 @@ def _fwd_kernel(
     tl.store(out_ptrs, acc, mask=offs_m[:, None] < cur_batch_seq_len)
 
 
-@wrap_jit_func
 @triton.jit
 def _fwd_split_kernel_quant(
     Q,
@@ -561,7 +555,6 @@ def _fwd_split_kernel_quant(
     tl.store(Acc_out + off_meta + 1 + tl.arange(0, 1), l_i)
 
 
-@wrap_jit_func
 @triton.jit
 def _fwd_kernel_quant(
     Q,
@@ -802,7 +795,6 @@ def alibi_paged_attention_fwd(
     grid = (batch, head, triton.cdiv(max_input_len, BLOCK))  # batch, head,
 
     num_warps = 4 if Lq <= 64 else 8
-    kernel_meta = get_kernel_meta(q)
     is_decoding = q.shape[-3] == b_seq_len.size(0)
     if not is_decoding:
         if quant_policy > 0:
@@ -846,8 +838,7 @@ def alibi_paged_attention_fwd(
                                     BLOCK_DMODEL=Lq,
                                     BLOCK_N=BLOCK,
                                     num_warps=num_warps,
-                                    num_stages=1,
-                                    **kernel_meta)
+                                    num_stages=1)
         else:
             _fwd_kernel[grid](q,
                               k,
@@ -880,8 +871,7 @@ def alibi_paged_attention_fwd(
                               BLOCK_DMODEL=Lq,
                               BLOCK_N=BLOCK,
                               num_warps=num_warps,
-                              num_stages=1,
-                              **kernel_meta)
+                              num_stages=1)
     else:
         SPLIT_K = 4
         grid = (batch, head, SPLIT_K)
@@ -927,8 +917,7 @@ def alibi_paged_attention_fwd(
                                           BLOCK_DMODEL=Lq,
                                           BLOCK_N=BLOCK,
                                           num_warps=4,
-                                          num_stages=1,
-                                          **kernel_meta)
+                                          num_stages=1)
 
         else:
             _fwd_split_kernel[grid](q,
@@ -961,8 +950,7 @@ def alibi_paged_attention_fwd(
                                     BLOCK_DMODEL=Lq,
                                     BLOCK_N=BLOCK,
                                     num_warps=4,
-                                    num_stages=1,
-                                    **kernel_meta)
+                                    num_stages=1)
 
         grid = (batch, head)
         _reduce_split_kernel[grid](acc,
@@ -977,5 +965,4 @@ def alibi_paged_attention_fwd(
                                    SPLIT_K=SPLIT_K,
                                    BLOCK_DMODEL=Lq,
                                    num_warps=num_warps,
-                                   num_stages=1,
-                                   **kernel_meta)
+                                   num_stages=1)
