+++ lmdeploy/pytorch/kernels/cuda/fused_lora.py
@@ -38,8 +38,6 @@ def _atomic_store(ptrs, val, mask):
     configs=get_autotune_config(),
     key=['N', 'K'],
     restore_value=['c_ptr'],
-    warmup=5,
-    rep=20,
 )
 @triton.jit
 def _fused_lora_kernel(
